WEBC-340: Restructuring services to split exception directory and put the exception to where they belong
