+++ src/Service/RequestHandler/Exception/InvalidRequestHandler.php
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
-namespace Shopgate\CloudIntegrationSdk\Service\Exception;
+namespace Shopgate\CloudIntegrationSdk\Service\RequestHandler\Exception;
 
 use Throwable;
